+++ CHANGELOG.md
@@ -3,5 +3,10 @@ All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
 
+## 2021.1 - 2021/09/07
+- Add `SimpleConf` to enable a user to extend the capabilities of FHConfParser
+  such that command-line arguments can be used to override config options.
+
+
 ## 2021 - 2021/08/21
 - First release